Introduce `Icinga\Module\Icingadb\Common\SearchControls`
